requirements.txt
pytest
requests
pydantic
python-dotenv
pytest-html